DXF Layer names added.

+++ developerDoc/buildDevDoc.py
@@ -10,17 +10,20 @@ print('**********************')
 print('Building Developer doc')
 print('**********************')
 print()
-print('Removing already existing html files to force a new build...')
 
 dirpath = os.path.dirname(os.path.realpath(__file__))
 
-pathName = os.path.join(dirpath, 'build/doctrees')
-if os.path.isdir(pathName):
-    shutil.rmtree(pathName)
+print('Delete old doc to force a complete build? [y/ n]')
+answ = input('Default= n ')
 
-pathName = os.path.join(dirpath, 'build/html')
-if os.path.isdir(pathName):
-    shutil.rmtree(pathName)
+if answ == 'y':
+    pathName = os.path.join(dirpath, 'build/doctrees')
+    if os.path.isdir(pathName):
+        shutil.rmtree(pathName)
+
+    pathName = os.path.join(dirpath, 'build/html')
+    if os.path.isdir(pathName):
+        shutil.rmtree(pathName)
 
 print()
 print('...starting Sphinx...')
